+++ homeassistant/components/lidarr/sensor.py
@@ -18,9 +18,8 @@ from homeassistant.const import UnitOfInformation
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import LidarrConfigEntry
 from .const import BYTE_SIZES
-from .coordinator import LidarrDataUpdateCoordinator, T
+from .coordinator import LidarrConfigEntry, LidarrDataUpdateCoordinator, T
 from .entity import LidarrEntity
